requirements.txt
# Project's Packages
tensorflow==2.16.2
tensorflow-metal==1.1.0; sys_platform == "darwin" and platform_machine == "arm64"
keras==3.3.3
tensorflow-datasets==4.9.4
numpy>=1.26.0,<2.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyyaml>=6.0.0
importlib_resources>=6.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
# api
fastapi
uvicorn
